requirements.txt
aiogram
deep_translator